bug: Fix stackoverflow on asset reload.
There is a way to create a circular dependency graph with asset loaders.
I don't know how I managed to do it. I have tried to create a minimal
example, but it has not exhibited the error yet. This commit has two
fixes: one at the recursion point, and one at the insertion point.

Check for self loops. Warn on self detection and do not loop.

Issue warning when an asset wants to mark itself as a dependency and do
not allow inserting itself as a dependent.